+++ src/complex/arith.rs
@@ -19,14 +19,12 @@ use {Complex, Float};
 use Integer;
 use complex::big::{Ordering2, Round2, ordering2, raw_round2};
 use ext::mpc as xmpc;
-use float::Special;
 use gmp_mpfr_sys::mpc::{self, mpc_t};
 use inner::{Inner, InnerMut};
 use ops::{AddAssignRound, AddFrom, AddFromRound, AssignRound, DivAssignRound,
           DivFrom, DivFromRound, MulAssignRound, MulFrom, MulFromRound,
           NegAssign, Pow, PowAssign, PowAssignRound, PowFrom, PowFromRound,
           SubAssignRound, SubFrom, SubFromRound};
-use std::iter::{Product, Sum};
 use std::ops::{Add, AddAssign, Div, DivAssign, Mul, MulAssign, Neg, Shl,
                ShlAssign, Shr, ShrAssign, Sub, SubAssign};
 use std::os::raw::c_int;
@@ -553,42 +551,9 @@ unsafe fn mul_sub(
     xmpc::mulsub(rop, (mul.lhs.inner(), mul.rhs.inner()), sub, rnd)
 }
 
-// use Complex::sum instead of fold! for iterator sum, to get correct rounding
-
-impl Sum for Complex {
-    #[inline]
-    fn sum<I>(mut iter: I) -> Complex
-    where
-        I: Iterator<Item = Complex>,
-    {
-        match iter.next() {
-            Some(first) => {
-                let others = iter.collect::<Vec<_>>();
-                first + Complex::sum(others.iter())
-            }
-            None => Complex::with_val(53, Special::Zero),
-        }
-    }
-}
-
-impl<'a> Sum<&'a Complex> for Complex {
-    #[inline]
-    fn sum<I>(mut iter: I) -> Complex
-    where
-        I: Iterator<Item = &'a Complex>,
-    {
-        match iter.next() {
-            Some(first) => first.clone() + Complex::sum(iter),
-            None => Complex::with_val(53, Special::Zero),
-        }
-    }
-}
-
-fold! { Complex, Product product, Complex::with_val(53, 1), Mul::mul }
-
 #[cfg(test)]
 mod tests {
-    use {Assign, Complex};
+    use Complex;
     use ops::Pow;
 
     #[test]
@@ -637,32 +602,5 @@ mod tests {
 
         assert_eq!(Complex::with_val(53, (&lhs).pow(ps)), lhs.clone().pow(ps));
         assert_eq!(Complex::with_val(53, (&lhs).pow(pd)), lhs.clone().pow(pd));
-    }
-
-    #[test]
-    fn check_sum() {
-        let nothing = Vec::<Complex>::new();
-        let empty1: Complex = nothing.iter().sum();
-        assert_eq!(empty1, 0);
-        assert_eq!(empty1.prec(), (53, 53));
-        let empty2: Complex = nothing.into_iter().sum();
-        assert_eq!(empty2, 0);
-
-        let values = vec![
-            Complex::with_val(4, (5.0, 1024.0)),
-            Complex::with_val(4, (1024.0, 15.0)),
-            Complex::with_val(4, (-1024.0, -1024.0)),
-            Complex::with_val(4, (-4.5, -16.0)),
-        ];
-        let sum1: Complex = values.iter().sum();
-        assert_eq!(sum1, (0.5, -1.0));
-        assert_eq!(sum1.prec(), (4, 4));
-        let mut sum2 = Complex::new(4);
-        sum2.assign(Complex::sum(values.iter()));
-        assert_eq!(sum2, (0.5, -1.0));
-        sum2 += Complex::sum(values.iter());
-        assert_eq!(sum2, (1.0, -2.0));
-        let sum3: Complex = values.into_iter().sum();
-        assert_eq!(sum3, (0.5, -1.0));
     }
 }